+++ src/_old-code/ui/add-notes-journal-modal.ts
@@ -47,7 +47,7 @@ export class AddNotesJournalModal extends Modal {
     private config: CalendarConfig,
   ) {
     super(app);
-    this.journal = this.manager.get(this.config.name) as CalendarJournal;
+    this.journal = this.manager.get(this.config.id) as CalendarJournal;
   }
 
   onOpen(): void {
@@ -361,6 +361,7 @@ export class AddNotesJournalModal extends Modal {
 
   private async processNote(note: TFile): Promise<void> {
     this.writeToConsole(`Processing ${note.basename}`);
+    if (note.extension !== "md") return this.writeToConsole(`Not a markdown file - skipped`);
     const journalData = await this.manager.getJournalData(note.path);
     this.console_ident++;
     try {